+++ admin/wcis-checkout.php
@@ -5,8 +5,8 @@
 class WCIS_Checkout {
 
   function __construct() {
-    // add_filter('woocommerce_checkout_fields', array($this, 'reorder_fields') );
-    add_filter('woocommerce_default_address_fields', array($this, 'reorder_fields') );
+    // TODO: this should be in theme, not in plugin
+    // add_filter('woocommerce_default_address_fields', array($this, 'reorder_fields') );
 
     add_action('woocommerce_checkout_update_user_meta', array($this, 'update_user_meta'), 99, 2);
     add_action('woocommerce_checkout_update_order_meta', array($this, 'update_order_meta'), 99, 2);
@@ -23,17 +23,6 @@ class WCIS_Checkout {
     @return array - The ordered list of fields
   */
   function reorder_fields($fields) {
-    // default priorities:
-    // 'first_name' - 10
-    // 'last_name' - 20
-    // 'company' - 30
-    // 'country' - 40
-    // 'address_1' - 50
-    // 'address_2' - 60
-    // 'city' - 70
-    // 'state' - 80
-    // 'postcode' - 90
-
     $fields['state']['priority'] = 42;
     $fields['city']['priority'] = 44;
     $fields['postcode']['priority'] = 46;